src/requirements.txt
requests==2.26.0
scapy==2.4.5
Pysocks==1.7.1
